+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.0.2 2014-xx-xx
+## 0.0.2 2014-01-20
 - Calls ejabberdctl load_config to reload ACLs when config changes. This means that ejabberd is no longer restarted when config changes.
 - Allow ejabberd config file to be specified
 - Allow package name and package ensure to be specified